pyproject.toml
[tool.poetry]
name = "taurius-installer"
version = "0.1.0"
description = ""
authors = ["Pedro Augusto Barbosa Aparecido <pedr.augustobarbosa.aparecido@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.12,<3.13"
appdata = "^2.2.1"
requests = "^2.31.0"
cefpython3 = "^66.1"
pythonnet = "^3.0.3"
pywebview = "^5.1"
flask = "^3.0.3"
flask-socketio = "^5.3.6"
python-dotenv = "^1.0.1"
flask-session = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
taskipy = "^1.13.0"
pytest-cov = "^5.0.0"

[tool.black]
line-length = 80
target-version = ["py312"]
exclude = '''
/(
  \.git
 | \.hg
 | \.mypy_cache
 | \.tox
 | venv
 | _build
 | buck-out
 | build
 | \.flake8
 )/
'''

[tool.isort]
profile = "black"
line_length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
lint-fix = "echo \"Formmating with black\" && black . && echo \"Formmating with isort\" && isort ."
test = "pytest -s -x -vv --cov=taurius_installer"
pre_test = "task lint"
post_test = "coverage html && py tasks/open_cov.py"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--ignore=tasks"
